requirements.txt
Flask==2.2.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
gunicorn==20.1.0
requests==2.31.0
stripe==7.11.0
psycopg2-binary==2.9.9
paypalrestsdk==1.13.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
flask-session==0.5.0
Werkzeug==2.2.3
bcrypt==4.1.2
supabase==1.0.3
Jinja2==3.1.2
MarkupSafe==2.1.2
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.32
python-dateutil==2.9.0
pytz==2024.1
itsdangerous==2.1.2
click==8.1.3